+++ Source/Core/VideoBackends/OGL/OGLPipeline.cpp
@@ -5,7 +5,6 @@
 
 #include "Common/Assert.h"
 
-#include "VideoBackends/OGL/OGLRender.h"
 #include "VideoBackends/OGL/OGLShader.h"
 #include "VideoBackends/OGL/OGLVertexManager.h"
 #include "VideoBackends/OGL/ProgramShaderCache.h"
