+++ Source/Core/VideoCommon/AbstractGfx.cpp
@@ -13,6 +13,8 @@
 #include "VideoCommon/VertexManagerBase.h"
 #include "VideoCommon/VideoConfig.h"
 
+std::unique_ptr<AbstractGfx> g_gfx;
+
 bool AbstractGfx::IsHeadless() const
 {
   return true;
