+++ Source/Core/VideoCommon/BoundingBox.cpp
@@ -13,6 +13,8 @@
 
 #include <algorithm>
 
+std::unique_ptr<BoundingBox> g_bounding_box;
+
 void BoundingBox::Enable(PixelShaderManager& pixel_shader_manager)
 {
   m_is_active = true;
